Merged from trunk 4796:4810
Add some ACCEPT_LICENSE docs.

Put the package.keywords docs in alphabetical order with the others.

Add some package.license docs.

Add news about ACCEPT_LICENSE and package.license.

Make config.load_infodir also load the data into the env layer in order
to ensure that it has preference over what may have been there.

Don't limit the size of USE in config.load_infodir.

Add package.license to the list of files.

Don't put -* in the env layer, since that layer isn't part of the
incremental stacking in regenerate.

In getmaskingstatus(), add missing check for * in acceptable_licenses.

Don't display the "Please update portage" message in --quiet mode.
Thanks to solar.

Remove paths from _eclasses_ serialization in the cache.  This makes the
mtimes of the eclasses the only thing that distinguishes them, but the
probablility of collision is negligible.  This same _eclasses_ format
will be used to serialize eclass mtimes in cache that is distributed via
the rsync mirrors.  The deserialization code can handle mixtures of both
the old and new formats.

Enable recognition of _mtime_ that doesn't rely on cache mtime mangling.

Add ACCEPT_LICENSE back to incrementals so that profiles can affect it.

Disable ACCEPT_LICENSE support until the details of license groups are
decided and implemented.


svn path=/main/branches/prefix/; revision=4812