Merged from trunk 4847:4855
Do not pass the "clean" parameter to doebuild since it causes ${T} to be
cleaned up when it shouldn't be.  Thanks to Mr_Bones_ for reporting.

Don't bother to insert /usr/lib/portage/pym into sys.path from inside a
module.  That's the caller's job.

Add --with-bdeps to emerge --help output.

Update help for --deep.

Update help for greedy behavior of --update.

Add backward compatibility code for cache modules that override has_key
instead of __contains__.


svn path=/main/branches/prefix/; revision=4857