Merged from trunk 4876:4881
For bug #153295, do a sanity check for non-string types in
config.environ().

For getmaskingstatus messages, don't assume ARCH is a valid keyword,
since that assumption is not necessarily correct for operating systems
other than Linux.  Thanks to Diego Petteno for reporting.

Just use stat instead of lstat for config protect, so that broken
symlinks are just ignored.


svn path=/main/branches/prefix/; revision=4883