Merged from trunk 4915:4929
For bug #153951, fail with a simple "Permission denied" error if the
merge phase doesn't have write access to $ROOT.

Silently omit info_pkgs that aren't installed.  Thanks to solar.

When transferring cache after sync, compare all metadata to make sure it
is identical (don't trust _mtime_ and _eclasses_ alone).  Local tests
show that this can add approximately 15% more time to the metadata
update, but it's needed for things like bug #153591.

Make sure trg is defined.

Fix --ignore-default-opts for bug #153980.

Revert implicit rdepend behavior back to the way it was in
portage-2.0.51.  Thanks to vapier for the patch.  The original change
seems to have been an  accidentally hunk (referencing bug #58819) that
was committed to cvs along with another patch in revision 1.201.2.32 of
ebuild.sh.


svn path=/main/branches/prefix/; revision=4939