Merged from trunk 4946:4959
Prevent the --newuse check in depgraph.create() that is intended for the
ebuild type from being done on binary packages (which can lead to a
KeyError).  The --newuse check for binaries is elsewhere (in
select_dep).

Ensure that sys-apps/portage is as close as possible to the beginning of
the merge list.  This is accomplished by merging it before any soft
deps that it may have.

Remove "update portage before proceeding" message that is no longer
relevant.

Remove old code the used to bump portage near the top of the merge list.
It's ineffective now that a complete depgraph is built.

Make sure the node is still in the graph so that a KeyError can't
happen.

For bug #154223, bridge new-style virtuals into old-style virtuals in
order to prevent old-style virtuals from making new-style virtuals
invisible.

Make config.add_new_virtuals() work properly even if self.virtuals has        
previously been initialized.

For bug #154220, run /etc/portage/bin/post_sync at the end of
emerge-webrsync.


svn path=/main/branches/prefix/; revision=4961