Merged from trunk 4998:5012
Thanks again to Chris White for these additional portage_dep docstrings.

Make digestcheck() fail if an ebuild isn't listed in the Manifest.
Thanks to pva for reporting.

Fix emerge --sync with a cvs:// uri so that it runs
/etc/portage/bin/post_sync  like it would with a normal rsync:// uri.
Thanks to drizzt for reporting.

Fix initial cvs checkout logic for bug #100478.

For initial gentoo-x86 checkout, if $PORTDIR exists and is empty then
remove it so that the gentoo-x86 checkout can replace it.

For bug #154554, validate all profile virtuals atoms.

Fix depgraph.pkg_node_map so that it properly supports $ROOT.



svn path=/main/branches/prefix/; revision=5014