Merged from trunk 5091:5107
Use mydbapi is None to eliminate the return_all_deps parameter from
dep_check and dep_zapdeps.  The result is the same and this way we
avoid having an additional parameter with overlapping meaning.

Allow dep_wordreduce to mark things satisfied by package.provided even
when mydbapi is None.

For depclean, just pass in None for mydbapi (instead of an empty
fakedbapi instance) since that's now supported.

Remove redundant package.provided handling that is now covered by          
dep_wordreduce.

Allow the ROOT setting to come from make.conf.  Thanks to Daniel
Barkalow <barkalow@iabervon.org> for this patch (submitted on the
gentoo-portage-dev mailing list).

For bug #155739, only check if a dep is in the args when myroot ==
self.target_root.

Print a warning message if a user requests a package from
package.provided (since it won't be merged).

Enable PORTAGE_CONFIGROOT to be set via an new --config-root option.
Thanks to Daniel Barkalow <barkalow@iabervon.org> for the initial patch
(submitted on the gentoo-portage-dev mailing list).


svn path=/main/branches/prefix/; revision=5108