Merged from trunk 5120:5123
For bug #150585, stack package.keywords incrementally so that -keyword
works as expected for any incremental variable.  This brings the
gvisible and getmaskingstatus logic into sync so that the confusing
(masked by: ) message will not be generated in cases where
package.keywords contains -keyword followed by keyword.

For bug #156014, always round up the next kB so that it doesn't show 0
kB when some small file still needs to be fetched.

Allow -keyword to be matched since some ebuilds use it.  Thanks to Brian
Harring reporting on bug #150585.


svn path=/main/branches/prefix/; revision=5124