Merged from trunk 5132:5135
Never overwrite an untouched config file and trigger config protection
instead (in case the new config has an unwanted change).  In cases where
an installed package in the same slot owns a protected file that will be
merged, bump the mtime on the installed file in order to ensure that it
isn't unmerged.  Thanks to Ciaran McCreesh <ciaranm@ciaranm.org> for his
suggestion on bug #8423, comment #40.

Update the dblink.isprotected() docstring.

For bug #8423, make etc-update preserve timestamps when possible.


svn path=/main/branches/prefix/; revision=5138