Merged from trunk 5149:5157
Always verify the ebuild checksums before executing it.  Thanks to solar
for the suggestion and antarus for the initial patch.

For manifest and digest phases, use a global variable to temporarily
exempt the depend phase from manifest checks (in cache regeneration is
triggered).

Temporarily exempt the depend phase during the help phase also.

Avoid checking the same Manifest several times in a row during a regen
with an empty cache.

For consistent behavior in all timezones, make emerge-webrsync use UTC
time for decisions about which snapshots to download.

Only attempt to verify the Manifest if the ebuild is actually in a
portage tree.

Make sure that *all* of the ebuilds are listed in the Manifest.

Make sure that all depend phases triggered by the digest phase are
exempt from digest checks.



svn path=/main/branches/prefix/; revision=5158