Merged from trunk 5158:5161
Skip digest checks when digest is in features.

Make portage.digestgen() automatically exempt the depend phase from
Manifest checks so that it works for repoman without triggering a digest
verification failure.

Move the global statement out of the try block so that it's clearly in
scope for the finally block.


svn path=/main/branches/prefix/; revision=5162