Merged from trunk 5261:5264
Don't collapse the tuple into a hash, since dict can use __eq__ to avoid
problems with hash collisions.  Thanks to Brian Harring for reporting.

For bug #157734, fix repoman isvalidatom() checks.  Thanks to Brian
Harring for the patch.

Add an allow_blockers parameter to isvalidatom so that the behavior is
optional.  Thanks to Brian Harring for the suggestion.


svn path=/main/branches/prefix/; revision=5265