Merged from trunk 5479:5485
Don't automatically fix digests in --pretend mode.
(prefix also has isSvn mode)

s/prefix/suffix/ Thanks to hydrogen for reporting on irc.

s/suffix/version prefix/ since it really is a prefix for the version
string.

For bug #160893, ignore a KeyError during cache cleansing since that key
isn't wanted anyway.

Implement QA checks that attempt to detect changed shopts and IFS in
ebuild/eclass/profile code.  Spit a warning if that is the case.  I
expect this check will eventually reset the shopts and IFS, but in this
case it may break things that depend on the current behavior. (WarnerBro)

Antarus' bash improvements, thanks for Genone for reporting this major
bug. (WarnerBro)



svn path=/main/branches/prefix/; revision=5500