Merged from trunk 5981:5990
For bug #166939, make --nodeps work like a normal package for new-style
virtuals.

Remove unnecessary import that triggers DEPRECATION NOTICE.

Fix package.provided breakage from the fix for bug #166939.

For bug #166785, check for invalid categories in the world file.

For bug #163990, warn about ebuilds with invalid categories and do not
allow them to be installed.

For bug #163990, warn about binary packages with invalid categories and
do not allow them to be installed.





svn path=/main/branches/prefix/; revision=6078