Merged from trunk 6152:6169
you can't insert into sys.path without importing sys... (WarnerBro)

In the same vein as copyright, add Id and svn:keywords to it (WarnerBro)

make sure uid is an int (WarnerBro)

Pass waiting_msg into recursive lockfile() calls.

Only prune duplicate consecutive nodes if they both have depth == 0.

For bug #169187, make --depclean prune unneeded slots such that it is
consistent with the world update algorithm.

Add a note about --with-bdeps in the --depclean failure message.

Fix the minimum index for pruning consecutive duplicate nodes.

Remove the KVERS environment variable since it's code that creates it is
broken anyway (only sets KVERS="." in the general case).

Display blockers at the end of the list so that it will more often be
easy for the user to see them in a terminal.  Thanks to iapitus and
hydrogen for the suggestion.




svn path=/main/branches/prefix/; revision=6172