Hi Grobian,
in prefix-portage/subst-install.in there is a problem using ${TMP-/tmp}
as temporary directory instead of ${TMPDIR-/tmp}.

TMP is undefined for unix, TMPDIR needs to be used instead.
But on Interix/Windows, TMP must be defined as dos-style path...

Thanks!

/haubi/

(so far for the smell of Windows, thanks Michael Haubenwallner)



svn path=/main/branches/prefix/; revision=6538