Bugs #168772 and #193695 - During unmerge, only ignore specific exceptions raised from unlink() and rmdir() calls.
Bug #193695 - Add support for FreeBSD chflags during unmerge. This
code is adapted from the code that already exists in movefile()
for the merge phase.

In movefile() FreeBSD chflags handling, use chflags instead of
lchflags when temporarily adjusting the flags on the parent
directory since we want to follow any symlinks to the real
parent directory.

(trunk r7803:7807)

svn path=/main/branches/2.1.2/; revision=7832