Intermediate commit.
Finished stage1 of chpathtool pimp-up.

The idea is that chpathtool copies a directory into another, doing its
job, retaining the files' properties, like cp -a (ok, cp -pPR) would do.
The implementation is now up to copying, setting the permission bits and
owner stuff.  TODO: the *times, and then it should be done.


svn path=/main/branches/prefix/; revision=8095