PYTHON_EPREFIX -> get rid of retarded name
First shot at some EAPI scheme which is closer towards the trunk than it
is now.  Main tree ebuilds can technically be marked EAPI=1 these days,
which is not reflected in prefix, as we use the static EAPI="prefix".
To gain the same EAPI levels here, EAPI in prefix is extended to consist
of two parts, "prefix" and the EAPI number of the main tree.  Of course
the latter can be omitted if it 0 and/or omitted in the main tree.


svn path=/main/branches/prefix/; revision=8455