Manual merge of r10325 from trunk.
The patch itself is a backport of a Prefix "feature" -- repoman SVN mode
-- but the backported patch was fixed/touched up/modified.  The changes
in this commit are those changes applied in the trunk.

I decided to remove any Prefix specific hacks that are dropped in the
trunk (e.g. avoiding a commit for the headers in SVN mode) to remain
fully compatible.


svn path=/main/branches/prefix/; revision=10341