Do a humble attempt to implement something along the lines of bug #229033
Big problem with SVN is that retrieving the known files is slow and
expensive (it goes over the network), hence should be avoided.

Imlemented Manifest commit in the first batch, if no headers change.
This implementation is a (hopefully) 90% coverage one.  First, not all
keywords are recognised (they never were), second we do not check in
SVN's case if the used keyword is the one that SVN expands.  What we
just try to achieve is to avoid the single Manifest commit on SVN repos
which by default do not use header expansion.  This objective seems to
be made with this commit.


svn path=/main/branches/prefix/; revision=11211