Original commit:
   | 11233   | Add support for an new EAPI="2_pre0" value so that people    |
   | zmedico | who want to test USE deps can set the EAPI to something      |
   |         | other than 0 or 1. We can support as many different          |
   |         | experimental EAPI values as we need, and drop support for    |
   |         | them when the final EAPI 2 is defined.                       |
   
Changed into a Prefix enabled version, which is more generic.  It build
further on the idea that multiple properties are being present in EAPI.
I split it up in "should have" properties ("prefix" for now) that must
be present, and "can have" properties, that may be present.  This
patch is not perfect (can have EAPI="prefix 2_pre0 1 0") but fixing that
requires more thinking (successing properties, conflicting
properties...)


svn path=/main/branches/prefix/; revision=11240