Context: http://archives.gentoo.org/gentoo-alt/msg_2d3dfb3888ac4dcd4e64e1539ef039ff.xml
Let Portage not be tricked by symlinks when determining whether to
remove a directory.  This solves the case where the user added symlink
in place of the original directory and Portage would unmerge the
symlink.  With this patch Portage detects it is not dealing with a
directory but a symlink.  I think this is what was originally intended
in this code.  zmedico should give it a thorough look before merging
back to trunk.


svn path=/main/branches/prefix/; revision=14248