Remove support for 'extend', 'remove', and 'intersect' sets.conf section attributes in SetConfig.getSetAtoms() since the current implementation does not meet user expectations, as mentioned in bug #253802, comment #5:
Package set operators currently operate on atoms, but what users really need
is for them to operate on the packages themselves. This will allow one set to
add or subtract packages from another even though the sets to not use the exact
same atoms to refer to the given packages.

I imagine the way this should be done is to create a mapping of atom -> package
for each set, perform the intersection using the packages, and then map the
package intersection back into a set of atoms. (trunk r14679)

svn path=/main/branches/2.1.7/; revision=14690