shebang check: find all executable things (including symlinks)
autoconf-wrapper package installs an executable script in usr/lib/misc
which is not in path, hence, the script is not prefixed.  However,
symlinks from usr/bin (autoconf, autoreconf, etc.) point to it, hence
the script should have been prefixed.  We now also check symlinks and
change whatever they point to.  Because this may mean the same file
(target) is evaluated multiple times (symlinks pointing to it) the sed
expression had to be made idempotent such that we don't end up with a
multiple of prefixes in the shebang.