Implement PORTAGE_BINHOST ssh:// protocol support by spawning ssh and using it to stream the Packages file on stdout.
This assumes that you have ssh-agent running, for authentication.
For now, you won't be able to use this unless you define your own
FETCHCOMMAND_SSH and RESUMECOMMAND_SSH values. It's trivial to do
this if you create a helper script that calls rsync after
transforming the URI into the host:/path format that rsync uses.