Bug #337178 - Fix slot conflict backtracking to mask the higher version if appropriate.
Generally, we mask the higher version since this solves common
cases in which a lower version is needed so that all dependencies
will be satisfied (bug #337178). However, if existing_node happens
to be installed then we mask that since this is a common case that
is triggered when --update is not enabled.