_unmerge_pkgfiles: improve out-of-prefix unmerging check
Bug #345633 is likely caused by a corrupted/invalid CONTENTS entry (from
previous buggy Portage releases).  It shows, however, that the unmerge
code wasn't protected well enough against "unmerging" files outside of
the Prefix.  All entries are evaluated now to see if they are within the
Prefix and a warning is issued when they are not, preventing any further
actions to be taken.