REQUIRED_USE: always show requirements
This fixes a case in which the user would receive a "masked by:
REQUIRED_USE violated" message without any more information about
the specific requirements. Now, unsatisfied REQUIRED_USE is never
treated as a mask, since it really is nothing like a mask. It's much
more like an unsatisfied USE dependency, so we need to handle it in
a similar way.