LinkageMapXCoff: fixes for preserve-libs on AIX
Drop code identical to and inheritable from LinkageMapELF.

Update code that is duplicated from LinkageMapELF.

Actually query information for /all/ preserved shared objects,
not the last one only.
Maybe the KeyError-problem is gone for AIX due to this fix,
even if that workaround was inherited from LinkageMapELF.

Deal with existing helper files in a binpkg created with 'quickpkg'.
Eventually improve performance when creating helper files.