repoman: implemented echangelog functionality
Instead of calling echangelog, which on its turn has to query the VCS
again, use the existing information on changes made to the current
directory, and update the ChangeLog from Python itself.
This avoids a call to echangelog, and avoids again retrieving the same
VCS information as repoman already did.  It makes repoman independent
from external tools it didn't install itself, and should be faster in
general.