sandbox: move away from separate macossandbox
Dropped prefixsandbox feature (protected at path level of $EPREFIX), to
align with master branch more.
Made macossandbox and macosusersandbox an execution mode of sandbox on
OSX platforms.
Documented that OS X sandbox is used, but that it is awfully slow (hence
sandbox is still not enabled by default for OSX (in Prefix)).